+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates `Breadcrumbs` component props to include `customLinkComponent` and `linkProps`.
 - Updates Banner component to allow for HTML content in the `content` prop when passed as a string.
 - Updates `Searchbar` component styles on mobile breakpoints